requirements.txt
apache-airflow==2.8.1
requests
beautifulsoup4
psycopg2-binary